+++ packages/eslint-plugin/src/rules/explicit-module-boundary-types.ts
@@ -224,14 +224,18 @@ export default util.createRule<Options, MessageIds>({
         return false;
       }
 
-      if (node.type === AST_NODE_TYPES.VariableDeclarator) {
+      if (
+        node.type === AST_NODE_TYPES.VariableDeclarator ||
+        node.type === AST_NODE_TYPES.FunctionDeclaration
+      ) {
         return (
-          node.id.type === AST_NODE_TYPES.Identifier &&
+          node.id?.type === AST_NODE_TYPES.Identifier &&
           options.allowedNames.includes(node.id.name)
         );
       } else if (
         node.type === AST_NODE_TYPES.MethodDefinition ||
-        node.type === AST_NODE_TYPES.TSAbstractMethodDefinition
+        node.type === AST_NODE_TYPES.TSAbstractMethodDefinition ||
+        (node.type === AST_NODE_TYPES.Property && node.method)
       ) {
         if (
           node.key.type === AST_NODE_TYPES.Literal &&
@@ -481,7 +485,7 @@ export default util.createRule<Options, MessageIds>({
       }
       checkedFunctions.add(node);
 
-      if (isAllowedName(node.parent) || ancestorHasReturnType(node)) {
+      if (isAllowedName(node) || ancestorHasReturnType(node)) {
         return;
       }