+++ packages/eslint-plugin/tests/rules/naming-convention.test.ts
@@ -18,10 +18,9 @@ const parserOptions = {
   project: './tsconfig.json',
 };
 
-const formatTestNames: Readonly<Record<
-  PredefinedFormatsString,
-  Record<'valid' | 'invalid', string[]>
->> = {
+const formatTestNames: Readonly<
+  Record<PredefinedFormatsString, Record<'valid' | 'invalid', string[]>>
+> = {
   camelCase: {
     valid: ['strictCamelCase', 'lower', 'camelCaseUNSTRICT'],
     invalid: ['snake_case', 'UPPER_CASE', 'UPPER', 'StrictPascalCase'],
@@ -1132,13 +1131,11 @@ ruleTester.run('naming-convention', rule, {
         {
           const camelCaseVar = 1;
           function camelCaseFunction() {}
-          declare function camelCaseDeclaredFunction() {
-          };
+          declare function camelCaseDeclaredFunction() {};
         }
         const PascalCaseVar = 1;
         function PascalCaseFunction() {}
-        declare function PascalCaseDeclaredFunction() {
-        };
+        declare function PascalCaseDeclaredFunction() {};
       `,
       options: [
         { selector: 'default', format: ['camelCase'] },
@@ -1449,6 +1446,31 @@ ruleTester.run('naming-convention', rule, {
             'accessor',
             'enumMember',
           ],
+          format: ['PascalCase'],
+        },
+      ],
+    },
+    {
+      code: `
+        const obj = {
+          Foo: 42,
+          Bar() {
+            return 42;
+          },
+        };
+      `,
+      parserOptions,
+      options: [
+        {
+          selector: 'memberLike',
+          format: ['camelCase'],
+        },
+        {
+          selector: 'property',
+          format: ['PascalCase'],
+        },
+        {
+          selector: 'method',
           format: ['PascalCase'],
         },
       ],
@@ -1966,8 +1988,7 @@ ruleTester.run('naming-convention', rule, {
       code: `
         const PascalCaseVar = 1;
         function PascalCaseFunction() {}
-        declare function PascalCaseDeclaredFunction() {
-        };
+        declare function PascalCaseDeclaredFunction() {};
       `,
       options: [
         { selector: 'default', format: ['snake_case'] },