+++ packages/scope-manager/src/analyze.ts
@@ -61,6 +61,12 @@ interface AnalyzeOptions {
    * The source type of the script.
    */
   sourceType?: 'script' | 'module';
+
+  /**
+   * Emit design-type metadata for decorated declarations in source.
+   * Defaults to `false`.
+   */
+  emitDecoratorMetadata?: boolean;
 }
 
 const DEFAULT_OPTIONS: Required<AnalyzeOptions> = {
@@ -72,6 +78,7 @@ const DEFAULT_OPTIONS: Required<AnalyzeOptions> = {
   jsxFragmentName: null,
   lib: ['es2018'],
   sourceType: 'script',
+  emitDecoratorMetadata: false,
 };
 
 function mapEcmaVersion(version: EcmaVersion | undefined): Lib {
@@ -106,6 +113,9 @@ function analyze(
       providedOptions?.jsxFragmentName ?? DEFAULT_OPTIONS.jsxFragmentName,
     sourceType: providedOptions?.sourceType ?? DEFAULT_OPTIONS.sourceType,
     lib: providedOptions?.lib ?? [mapEcmaVersion(ecmaVersion)],
+    emitDecoratorMetadata:
+      providedOptions?.emitDecoratorMetadata ??
+      DEFAULT_OPTIONS.emitDecoratorMetadata,
   };
 
   // ensure the option is lower cased