+++ packages/shared-fixtures/fixtures/jsx/escape-patters-multi.src.js
@@ -1,0 +1,1 @@
+<a href="test=1&amp;bar=1&#123;">&nbsp; test</a>;