+++ packages/typescript-estree/tests/lib/persistentParse.test.ts
@@ -116,7 +116,7 @@ function baseTests(
 
   it('allows parsing of deeply nested new files', () => {
     const PROJECT_DIR = setup(tsConfigIncludeAll, false);
-    const bazSlashBar = path.join('baz', 'bar') as 'baz/bar';
+    const bazSlashBar = 'baz/bar' as const;
 
     // parse once to: assert the config as correct, and to make sure the program is setup
     expect(() => parseFile('foo', PROJECT_DIR)).not.toThrow();
@@ -141,7 +141,7 @@ function baseTests(
     fs.mkdirSync(path.join(PROJECT_DIR, 'src', 'bat'));
     fs.mkdirSync(path.join(PROJECT_DIR, 'src', 'bat', 'baz'));
 
-    const bazSlashBar = path.join('bat', 'baz', 'bar') as 'bat/baz/bar';
+    const bazSlashBar = 'bat/baz/bar' as const;
 
     // write a new file and attempt to parse it
     writeFile(PROJECT_DIR, bazSlashBar);
@@ -151,7 +151,7 @@ function baseTests(
 
   it('allows renaming of files', () => {
     const PROJECT_DIR = setup(tsConfigIncludeAll, true);
-    const bazSlashBar = path.join('baz', 'bar') as 'baz/bar';
+    const bazSlashBar = 'baz/bar' as const;
 
     // parse once to: assert the config as correct, and to make sure the program is setup
     expect(() => parseFile('foo', PROJECT_DIR)).not.toThrow();
@@ -279,7 +279,7 @@ describe('persistent parse', () => {
 
     it('handles tsconfigs with no includes/excludes (nested)', () => {
       const PROJECT_DIR = setup({}, false);
-      const bazSlashBar = path.join('baz', 'bar') as 'baz/bar';
+      const bazSlashBar = 'baz/bar' as const;
 
       // parse once to: assert the config as correct, and to make sure the program is setup
       expect(() => parseFile('foo', PROJECT_DIR)).not.toThrow();