+++ packages/typescript-estree/tests/lib/semantic-diagnostics-enabled.test.ts
@@ -1,21 +1,23 @@
 import { readFileSync } from 'fs';
 import glob from 'glob';
+import path from 'path';
 import * as parser from '../../src';
-import { extname } from 'path';
 import { formatSnapshotName, isJSXFileType } from '../../tools/test-utils';
 
 /**
  * Process all fixtures, we will only snapshot the ones that have semantic errors
  * which are ignored by default parsing logic.
  */
-const FIXTURES_DIR =
-  '../../node_modules/@typescript-eslint/shared-fixtures/fixtures';
-const testFiles = glob.sync(`${FIXTURES_DIR}/**/*.src.*`);
+const FIXTURES_DIR = path.join(__dirname, '../../../shared-fixtures/fixtures');
+
+const testFiles = glob.sync(`**/*.src.*`, {
+  cwd: FIXTURES_DIR,
+});
 
 describe('Parse all fixtures with "errorOnTypeScriptSyntacticAndSemanticIssues" enabled', () => {
   testFiles.forEach(filename => {
-    const code = readFileSync(filename, 'utf8');
-    const fileExtension = extname(filename);
+    const code = readFileSync(path.join(FIXTURES_DIR, filename), 'utf8');
+    const fileExtension = path.extname(filename);
     const config: parser.TSESTreeOptions = {
       loc: true,
       range: true,